[refactor] extract matchTeamLogo to own file
