+++ src/sections/polymarket/polymarket-market-list.tsx
@@ -14,7 +14,7 @@ import { useTranslate } from 'src/locales'
 import PolymarketMarketToolbar from './polymarket-market-toolbar'
 import PolymarketMarketCard from './polymarket-market-card'
 import PolymarketEventCard from './polymarket-event-card'
-import { matchTeamLogo } from './polymarket-team-logos'
+import { matchTeamLogo } from './polymarket-team-match'
 
 import type { IPolymarketEvent } from 'src/types/polymarket'
