+++ src/sections/polymarket/view/polymarket-event-detail-view.tsx
@@ -24,7 +24,8 @@ import { useSettingsContext } from 'src/components/settings'
 import { fNumber } from 'src/utils/format-number'
 
 import PolymarketMarketCard from '../polymarket-market-card'
-import PolymarketTeamLogos, { matchTeamLogo } from '../polymarket-team-logos'
+import PolymarketTeamLogos from '../polymarket-team-logos'
+import { matchTeamLogo } from '../polymarket-team-match'
 import PolymarketTermsOverlay from '../polymarket-terms-overlay'
 import { sortMarketOptions } from '../polymarket-market-sort'
 import { usePolymarketAccountStatus } from '../use-polymarket-account-status'
